moved backend modules into modules directory
