+++ .gitignore
@@ -214,3 +214,5 @@ __marimo__/
 
 .gemini/
 gha-creds-*.json
+
+/infra/init-scripts/05-seed.sql
