+++ backend/src/modules/cpu/cpu.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 
-import { DatabaseModule } from '../database/database.module';
+import { DatabaseModule } from '../../database/database.module';
 import { CpuController } from './cpu.controller';
 import { CpuService } from './cpu.service';
 import { TypeOrmCpuRepository } from './gpu.typeorm.repository';
