+++ backend/src/modules/cpu/cpu.service.ts
