+++ backend/src/modules/cpu/entities/cpu.entity.ts
