+++ backend/src/modules/gpu/entities/gpu.entity.ts
